+++ test/utils/ArraysTest.js
@@ -184,9 +184,9 @@ suite("Arrays", function ArraysTest() {
         var ary1 = [1];
         var ary0 = [];
         var aryNull = null;
-        asr.equal(Arrays.hasItems(ary1), true);
-        asr.equal(Arrays.hasItems(ary0), false);
-        asr.equal(Arrays.hasItems(aryNull), false);
+        asr.equal(Arrays.hasAny(ary1), true);
+        asr.equal(Arrays.hasAny(ary0), false);
+        asr.equal(Arrays.hasAny(aryNull), false);
     });
     test("indexOfProp", function indexOfPropTest() {
         var ary = [{ id: "B" }, { id: "D" }, { id: "D" }, { id: "F" }];